# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PetSpace",
        "description": "Documentation for PetSpace Endpoints",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pet/allergen-dictionary": {
            "get": {
                "tags": [
                    "PetHealth"
                ],
                "summary": "Get allergen dictionary list",
                "description": "Retrieves a list of allergens and their classifications",
                "operationId": "getAllergenDictionary",
                "responses": {
                    "200": {
                        "description": "Allergy List retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Allergy List retrieved successfully"
                                        },
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID of the allergen",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "allergen": {
                                                        "description": "Name of the allergen",
                                                        "type": "string",
                                                        "example": "Beef"
                                                    },
                                                    "classification": {
                                                        "description": "Classification of the allergen",
                                                        "type": "string",
                                                        "example": "Food"
                                                    },
                                                    "species_affected": {
                                                        "description": "Species affected by the allergen",
                                                        "type": "string",
                                                        "example": "Dog & Cat"
                                                    },
                                                    "created_at": {
                                                        "description": "Timestamp of allergen creation",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "updated_at": {
                                                        "description": "Timestamp of allergen update",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Allergens not found"
                    }
                }
            }
        },
        "pet/{petId}/allergy/add/{allergenId}": {
            "post": {
                "tags": [
                    "PetHealth"
                ],
                "summary": "Add a pet allergy",
                "description": "Adds an allergy to a pet and returns the updated list of allergies",
                "operationId": "addPetAllergy",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "description": "ID of the pet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "allergenId",
                        "in": "path",
                        "description": "ID of the allergen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Allergy added successfully with updated allergy list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pet Allergy added successfully"
                                        },
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Record ID of the pet allergy",
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "allergen_id": {
                                                        "description": "ID of the allergen",
                                                        "type": "integer",
                                                        "example": 11
                                                    },
                                                    "allergen": {
                                                        "description": "Name of the allergen",
                                                        "type": "string",
                                                        "example": "Turkey"
                                                    },
                                                    "classification": {
                                                        "description": "Classification of the allergen",
                                                        "type": "string",
                                                        "example": "Food"
                                                    },
                                                    "added_by_first_name": {
                                                        "description": "First name of the person who added the allergy",
                                                        "type": "string",
                                                        "example": "Steve"
                                                    },
                                                    "added_by_last_name": {
                                                        "description": "Last name of the person who added the allergy",
                                                        "type": "string",
                                                        "example": "Tester"
                                                    },
                                                    "added_by_role": {
                                                        "description": "Role of the person who added the allergy, if applicable",
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            }
        },
...
        "/pet/{petId}/allergies": {
            "get": {
                "tags": [
                    "PetHealth"
                ],
                "summary": "Get pet allergy list",
                "description": "Retrieves a list of allergies for the specified pet",
                "operationId": "getPetAllergies",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "description": "ID of the pet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet allergy list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pet allergy list retrieved successfully"
                                        },
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Record ID of the pet allergy",
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "allergen_id": {
                                                        "description": "ID of the allergen",
                                                        "type": "integer",
                                                        "example": 11
                                                    },
                                                    "allergen": {
                                                        "description": "Name of the allergen",
                                                        "type": "string",
                                                        "example": "Turkey"
                                                    },
                                                    "classification": {
                                                        "description": "Classification of the allergen",
                                                        "type": "string",
                                                        "example": "Food"
                                                    },
                                                    "added_by_first_name": {
                                                        "description": "First name of the person who added the allergy",
                                                        "type": "string",
                                                        "example": "Steve"
                                                    },
                                                    "added_by_last_name": {
                                                        "description": "Last name of the person who added the allergy",
                                                        "type": "string",
                                                        "example": "Tester"
                                                    },
                                                    "added_by_role": {
                                                        "description": "Role of the person who added the allergy, if applicable",
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                }
            }
        },
...
    },
...
}